--- helpers/components/setID.js
@@ -1,6 +1,0 @@
-'use strict';
-module.exports = function (str) {
-  this.id = str; // set string into current context
-
-  return '';
-};
